@@ -30,7 +30,6 @@
   },
   "dependencies": {
     "@blueprintjs/core": "^4.0.0-beta.0",
-    "@blueprintjs/datetime": "^4.0.0-beta.0",
     "@codemirror/basic-setup": "^0.19.0",
     "@codemirror/commands": "^0.19.5",
     "@codemirror/gutter": "^0.19.9",
@@ -70,17 +69,17 @@
     "lorem-ipsum": "^2.0.4",
     "math-expressions": "^2.0.0-alpha59",
     "nanoid": "^3.1.31",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.3",
     "react-datetime": "^3.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^11.4.2",
     "react-is": "^17.0.1",
-    "react-mathquill": "^1.0.1",
+    "react-mathquill": "^1.0.3",
     "react-measure": "^2.5.2",
     "react-router": "^6.9.0",
     "react-router-dom": "^6.9.0",
-    "react-select": "^4.3.1",
+    "react-select": "^5.7.0",
     "react-table": "^7.7.0",
     "react-use-measure": "^2.1.1",
     "react-visibility-sensor-v2": "^1.0.0",
