@@ -1,6 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import { Route, Router, Switch } from 'react-router';
+import { createRoot } from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
 import ToolRoot from '@ToolRoot';
 
@@ -8,7 +7,8 @@ import Chat from './Chat';
 import { MathJaxContext } from 'better-react-mathjax';
 import { mathjaxConfig } from '../../Core/utils/math';
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <RecoilRoot>
     <MathJaxContext
       version={2}
@@ -17,6 +17,5 @@ ReactDOM.render(
     >
       <ToolRoot tool={<Chat key={'BaseTool'} />} />
     </MathJaxContext>
-  </RecoilRoot>,
-  document.getElementById('root'),
+  </RecoilRoot>
 );
