@@ -1,6 +1,6 @@
 import { MathJaxContext } from 'better-react-mathjax';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { mathjaxConfig } from '../../Core/utils/math.js';
 
 import UIDocs from './uiDocs.jsx';
@@ -8,7 +8,8 @@ import UIDocs from './uiDocs.jsx';
 // import { initialize } from '../imports/courseInfo';
 // initialize();
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <MathJaxContext
     version={2}
     config={mathjaxConfig}
@@ -16,4 +17,4 @@ ReactDOM.render(
   >
     <UIDocs />
   </MathJaxContext>
-  , document.getElementById('root')); 
+); 
